requirements.txt
aioodbc==0.5.0
asgiref==3.8.1
Bottleneck @ file:///C:/b/abs_816hr2khp1/croot/bottleneck_1731058648110/work
Brotli @ file:///C:/b/abs_c415aux9ra/croot/brotli-split_1736182803933/work
brotlipy==0.7.0
bson==0.5.10
certifi @ file:///home/conda/feedstock_root/build_artifacts/certifi_1754231422783/work/certifi
cffi @ file:///C:/b/abs_29_b57if3f/croot/cffi_1736184144340/work
cftime==1.6.4.post1
charset-normalizer @ file:///croot/charset-normalizer_1721748349566/work
click @ file:///C:/b/abs_42rbjngfys/croot/click_1744271606983/work
colorama @ file:///C:/b/abs_a9ozq0l032/croot/colorama_1672387194846/work
conda==22.11.1
conda-package-handling @ file:///C:/b/abs_7fz3aferfv/croot/conda-package-handling_1731369038903/work
conda_package_streaming @ file:///C:/b/abs_bdz9vbvbh2/croot/conda-package-streaming_1731366449946/work
cryptography @ file:///C:/b/abs_ae5lrkywtg/croot/cryptography_1740577842775/work
Django==4.2
django-cors-headers==4.7.0
django-filter==24.3
django-vite==3.1.0
djangorestframework==3.15.2
djangorestframework-gis==1.1
dnspython==2.7.0
exceptiongroup==1.2.2
Faker==33.1.0
Flask @ file:///home/conda/feedstock_root/build_artifacts/flask_1660054669620/work
GDAL @ file:///C:/b/abs_60gjdf6wzx/croot/gdal-split_1734448199581/work/build/swig/python
georinex==1.16.2
greenlet==3.2.2
gunicorn==23.0.0
hatanaka==2.8.1
idna @ file:///C:/b/abs_aad84bnnw5/croot/idna_1714398896795/work
importlib_metadata @ file:///C:/b/abs_9d0lp1ysmt/croot/importlib_metadata-suite_1732633510972/work
importlib_resources==6.5.2
iniconfig==2.0.0
itsdangerous @ file:///C:/b/abs_c4vwgdr5yn/croot/itsdangerous_1716533399914/work
Jinja2 @ file:///C:/b/abs_920kup4e6u/croot/jinja2_1741711580669/work
leaflet==0.0.3
MarkupSafe @ file:///C:/b/abs_a0ma7ge0jc/croot/markupsafe_1738584052792/work
menuinst @ file:///C:/ci/menuinst_1631733438520/work
mkl-service==2.4.0
mkl_fft @ file:///C:/Users/dev-admin/mkl/mkl_fft_1730823082242/work
mkl_random @ file:///C:/Users/dev-admin/mkl/mkl_random_1730822522280/work
model-bakery==1.20.0
ncompress==1.0.2
netCDF4==1.7.2
numexpr @ file:///C:/b/abs_05o8p7bfml/croot/numexpr_1730215959182/work
numpy @ file:///C:/b/abs_c1ywpu18ar/croot/numpy_and_numpy_base_1708638681471/work/dist/numpy-1.26.4-cp39-cp39-win_amd64.whl#sha256=f284633067fde4bb078a09085a1c9e99a3218c129f5312b7b3d76c9340442a25
packaging==24.1
pandas @ file:///C:/b/abs_e18w9ay4we/croot/pandas_1732735115740/work/dist/pandas-2.2.3-cp39-cp39-win_amd64.whl#sha256=f3b41dc9b2687250cbbb5587843986d13008da63bfb026b34d05d4d2f3eab666
pillow==11.0.0
pluggy @ file:///C:/b/abs_dfec_m79vo/croot/pluggy_1733170145382/work
psycopg2-binary==2.9.10
pycosat @ file:///C:/b/abs_18nblzzn70/croot/pycosat_1736868434419/work
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pymongo==4.13.0
pyodbc==5.2.0
pyOpenSSL @ file:///C:/b/abs_6cmmt1gx_z/croot/pyopenssl_1741343840223/work
pyproj==3.6.1
PySocks @ file:///C:/ci/pysocks_1605307512533/work
pytest==8.3.3
pytest-django==4.8.0
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/python-dateutil_1733215673016/work
pytz @ file:///home/conda/feedstock_root/build_artifacts/pytz_1742920838005/work
pywin32==308
requests @ file:///C:/b/abs_c3508vg8ez/croot/requests_1731000584867/work
ruamel-yaml-conda @ file:///C:/ci/ruamel_yaml_1616016898638/work
ruamel.yaml @ file:///C:/b/abs_30ee5qbthd/croot/ruamel.yaml_1666304562000/work
ruamel.yaml.clib @ file:///C:/b/abs_475flulhy_/croot/ruamel.yaml.clib_1745937166245/work
six @ file:///tmp/build/80754af9/six_1644875935023/work
SQLAlchemy==2.0.41
sqlparse==0.5.1
tomli==2.0.2
toolz @ file:///C:/b/abs_57avamxh36/croot/toolz_1736796279321/work
tqdm @ file:///C:/b/abs_0eh9b6xugj/croot/tqdm_1738945553987/work
typing_extensions @ file:///C:/b/abs_0ffjxtihug/croot/typing_extensions_1734714875646/work
tzdata==2024.1
urllib3 @ file:///C:/b/abs_7bst06lizn/croot/urllib3_1737133657081/work
Werkzeug @ file:///C:/b/abs_c7bupijx2_/croot/werkzeug_1737448709012/work
whitenoise==6.10.0
win-inet-pton @ file:///C:/ci/win_inet_pton_1605306162074/work
wincertstore==0.2
xarray @ file:///home/conda/feedstock_root/build_artifacts/xarray_1722348170975/work
zipp @ file:///C:/b/abs_2fcpcg89my/croot/zipp_1732630746332/work
zstandard @ file:///C:/b/abs_31t8xmrv_h/croot/zstandard_1731356578015/work
